requirements.txt
python-dotenv
requests
supabase